+++ react/constants/filterConstants.js
@@ -5,6 +5,14 @@ export const FACETS_RENDER_THRESHOLD = 10
  * Maps filter names to their i18n message keys
  */
 export const toggleFiltersValue = {
-  'next-day': 'store/search.filter.dynamic-estimate.next-day.name',
   'same-day': 'store/search.filter.dynamic-estimate.same-day.name',
+  'next-day': 'store/search.filter.dynamic-estimate.next-day.name',
+  'delivery-same-day':
+    'store/search.filter.dynamic-estimate.delivery-same-day.name',
+  'delivery-next-day':
+    'store/search.filter.dynamic-estimate.delivery-next-day.name',
+  'pickup-same-day':
+    'store/search.filter.dynamic-estimate.pickup-same-day.name',
+  'pickup-next-day':
+    'store/search.filter.dynamic-estimate.pickup-next-day.name',
 }