+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Changed
+
+- `dynamic-estimate` filter names to respect delivery method
+
 ## [3.141.1] - 2025-12-03
 
 ### Added